Allow root level iterables
