+++ html-common/src/html_tree/html_block.rs
@@ -1,13 +1,12 @@
+use super::html_iterable::HtmlIterable;
 use super::html_text::HtmlText;
 use crate::Peek;
-use boolinator::Boolinator;
 use proc_macro2::{Delimiter, TokenStream};
 use quote::{quote, quote_spanned, ToTokens};
 use syn::braced;
 use syn::buffer::Cursor;
 use syn::parse::{Parse, ParseStream, Result as ParseResult};
 use syn::token;
-use syn::Token;
 
 pub struct HtmlBlock {
     content: BlockContent,
@@ -54,39 +53,5 @@ impl ToTokens for HtmlBlock {
         };
 
         tokens.extend(quote_spanned! {brace.span=> #new_tokens});
-    }
-}
-
-struct HtmlIterable(TokenStream);
-
-impl Peek<()> for HtmlIterable {
-    fn peek(cursor: Cursor) -> Option<()> {
-        let (ident, _) = cursor.ident()?;
-        (ident.to_string() == "for").as_option()
-    }
-}
-
-impl Parse for HtmlIterable {
-    fn parse(input: ParseStream) -> ParseResult<Self> {
-        input.parse::<Token![for]>()?;
-        Ok(HtmlIterable(input.parse()?))
-    }
-}
-
-impl ToTokens for HtmlIterable {
-    fn to_tokens(&self, tokens: &mut proc_macro2::TokenStream) {
-        let stream = &self.0;
-        let new_tokens = quote! {
-            {
-                let mut __yew_vlist = ::yew::virtual_dom::VList::new();
-                for __yew_node in {#stream} {
-                    let __yew_vnode = ::yew::virtual_dom::VNode::from(__yew_node);
-                    __yew_vlist.add_child(__yew_vnode);
-                }
-                ::yew::virtual_dom::VNode::from(__yew_vlist)
-            }
-        };
-
-        tokens.extend(new_tokens);
     }
 }
