+++ html/tests/cases.rs
@@ -5,6 +5,9 @@ fn tests() {
     t.pass("tests/html-block-pass.rs");
     t.compile_fail("tests/html-block-fail.rs");
 
+    t.pass("tests/html-iterable-pass.rs");
+    t.compile_fail("tests/html-iterable-fail.rs");
+
     t.pass("tests/html-list-pass.rs");
     t.compile_fail("tests/html-list-fail.rs");
