+++ html/tests/html-block-fail.stderr
@@ -13,19 +13,18 @@ error[E0277]: `()` doesn't implement `std::fmt::Display`
    = note: required by `std::convert::From::from`
    = note: this error originates in a macro outside of the current crate (in Nightly builds, run with -Z external-macro-backtrace for more info)
 
-error[E0277]: `()` doesn't implement `std::fmt::Display`
+error[E0271]: type mismatch resolving `<[closure@$DIR/tests/html-block-fail.rs:19:28: 19:42 not_tree:_] as std::ops::FnOnce<({integer},)>>::Output == yew::virtual_dom::vnode::VNode<_>`
   --> $DIR/html-block-fail.rs:18:5
    |
 18 | /     html! {
 19 | |         <>{ for (0..3).map(|_| not_tree()) }</>
 20 | |     }
-   | |_____^ `()` cannot be formatted with the default formatter
+   | |_____^ expected (), found enum `yew::virtual_dom::vnode::VNode`
    |
-   = help: the trait `std::fmt::Display` is not implemented for `()`
-   = note: in format strings you may be able to use `{:?}` (or {:#?} for pretty-print) instead
-   = note: required because of the requirements on the impl of `std::string::ToString` for `()`
-   = note: required because of the requirements on the impl of `std::convert::From<()>` for `yew::virtual_dom::vnode::VNode<_>`
-   = note: required by `std::convert::From::from`
+   = note: expected type `()`
+              found type `yew::virtual_dom::vnode::VNode<_>`
+   = note: required because of the requirements on the impl of `std::iter::Iterator` for `std::iter::Map<std::ops::Range<{integer}>, [closure@$DIR/tests/html-block-fail.rs:19:28: 19:42 not_tree:_]>`
+   = note: required for the cast to the object type `dyn std::iter::Iterator<Item=yew::virtual_dom::vnode::VNode<_>>`
    = note: this error originates in a macro outside of the current crate (in Nightly builds, run with -Z external-macro-backtrace for more info)
 
 error[E0277]: `()` doesn't implement `std::fmt::Display`
@@ -43,4 +42,5 @@ error[E0277]: `()` doesn't implement `std::fmt::Display`
   = note: required by `std::convert::From::from`
   = note: this error originates in a macro outside of the current crate (in Nightly builds, run with -Z external-macro-backtrace for more info)
 
-For more information about this error, try `rustc --explain E0277`.
+Some errors occurred: E0271, E0277.
+For more information about an error, try `rustc --explain E0271`.
