+++ html/tests/html-iterable-fail.stderr
@@ -1,0 +1,33 @@
+error[E0277]: `()` is not an iterator
+ --> $DIR/html-iterable-fail.rs:3:1
+  |
+3 | test_html! { |t1| for }
+  | ^^^^^^^^^^^^^^^^^^^^^^^ `()` is not an iterator
+  |
+  = help: the trait `std::iter::Iterator` is not implemented for `()`
+  = note: required for the cast to the object type `dyn std::iter::Iterator<Item=yew::virtual_dom::vnode::VNode<_>>`
+  = note: this error originates in a macro outside of the current crate (in Nightly builds, run with -Z external-macro-backtrace for more info)
+
+error[E0277]: `()` is not an iterator
+ --> $DIR/html-iterable-fail.rs:4:1
+  |
+4 | test_html! { |t2| for () }
+  | ^^^^^^^^^^^^^^^^^^^^^^^^^^ `()` is not an iterator
+  |
+  = help: the trait `std::iter::Iterator` is not implemented for `()`
+  = note: required for the cast to the object type `dyn std::iter::Iterator<Item=yew::virtual_dom::vnode::VNode<_>>`
+  = note: this error originates in a macro outside of the current crate (in Nightly builds, run with -Z external-macro-backtrace for more info)
+
+error[E0271]: type mismatch resolving `<std::slice::Iter<'_, yew::virtual_dom::vnode::VNode<t3::TestComponent>> as std::iter::Iterator>::Item == yew::virtual_dom::vnode::VNode<_>`
+  --> $DIR/html-iterable-fail.rs:10:5
+   |
+10 |     html! { for empty.iter() }
+   |     ^^^^^^^^^^^^^^^^^^^^^^^^^^ expected reference, found enum `yew::virtual_dom::vnode::VNode`
+   |
+   = note: expected type `&yew::virtual_dom::vnode::VNode<t3::TestComponent>`
+              found type `yew::virtual_dom::vnode::VNode<_>`
+   = note: required for the cast to the object type `dyn std::iter::Iterator<Item=yew::virtual_dom::vnode::VNode<_>>`
+   = note: this error originates in a macro outside of the current crate (in Nightly builds, run with -Z external-macro-backtrace for more info)
+
+Some errors occurred: E0271, E0277.
+For more information about an error, try `rustc --explain E0271`.
